clean up MI app, remove csv stuff
